feat(admin): moved page menu items to the topbar
fixes #14045
