+++ engine/classes/Elgg/Menus/AdminHeader.php
@@ -3,6 +3,7 @@
 namespace Elgg\Menus;
 
 use Elgg\Menu\MenuItems;
+use Elgg\Menu\PreparedMenu;
 
 /**
  * Register menu items for the admin_header menu
@@ -37,6 +38,7 @@ class AdminHeader {
 			'href' => elgg_get_site_url(),
 			'parent_name' => 'account',
 			'priority' => 100,
+			'section' => 'alt',
 		]);
 		
 		// logout action
@@ -47,6 +49,7 @@ class AdminHeader {
 			'href' => elgg_generate_action_url('logout'),
 			'parent_name' => 'account',
 			'priority' => 1000,
+			'section' => 'alt',
 		]);
 		
 		// link to admin profile
@@ -59,7 +62,7 @@ class AdminHeader {
 				'alt' => $admin->getDisplayName(),
 			]),
 			'link_class' => 'elgg-avatar-small',
-			'icon_alt' => 'angle-down',
+			'section' => 'alt',
 		]);
 		
 		return $return;
@@ -86,10 +89,303 @@ class AdminHeader {
 		
 		$return[] = \ElggMenuItem::factory([
 			'name' => 'maintenance',
+			'icon' => 'warning',
 			'text' => elgg_echo('admin:configure_utilities:maintenance'),
 			'href' => 'admin/configure_utilities/maintenance',
 			'link_class' => 'elgg-maintenance-mode-warning',
 			'priority' => 700,
+		]);
+		
+		return $return;
+	}
+	
+	/**
+	 * Add the administer section to the admin header menu
+	 *
+	 * @param \Elgg\Event $event 'register', 'menu:admin_header'
+	 *
+	 * @return void|MenuItems
+	 */
+	public static function registerAdminAdminister(\Elgg\Event $event) {
+		if (!elgg_in_context('admin') || !elgg_is_admin_logged_in()) {
+			return;
+		}
+		
+		/* @var $return MenuItems */
+		$return = $event->getValue();
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'administer',
+			'text' => elgg_echo('menu:page:header:administer'),
+			'href' => false,
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'dashboard',
+			'text' => elgg_echo('admin:dashboard'),
+			'href' => 'admin',
+			'priority' => 10,
+			'parent_name' => 'administer',
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'plugins',
+			'text' => elgg_echo('admin:plugins'),
+			'href' => elgg_generate_url('admin', ['segments' => 'plugins']),
+			'priority' => 30,
+			'parent_name' => 'administer',
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'users',
+			'text' => elgg_echo('admin:users'),
+			'href' => 'admin/users',
+			'priority' => 40,
+			'parent_name' => 'administer',
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'upgrades',
+			'text' => elgg_echo('admin:upgrades'),
+			'href' => 'admin/upgrades',
+			'priority' => 600,
+			'parent_name' => 'administer',
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'administer_utilities',
+			'text' => elgg_echo('admin:administer_utilities'),
+			'href' => false,
+			'priority' => 50,
+			'parent_name' => 'administer',
+			'show_with_empty_children' => false,
+		]);
+		
+		return $return;
+	}
+	
+	/**
+	 * Prepare the users menu item in the administer section on admin pages
+	 *
+	 * @param \Elgg\Event $event 'prepare', 'menu:admin_header'
+	 *
+	 * @return PreparedMenu|null
+	 */
+	public static function prepareAdminAdministerUsersChildren(\Elgg\Event $event): ?PreparedMenu {
+		if (!elgg_in_context('admin') || !elgg_is_admin_logged_in()) {
+			return null;
+		}
+		
+		/* @var $result PreparedMenu */
+		$result = $event->getValue();
+		
+		$default = $result->getSection('default');
+		
+		/* @var $administer \ElggMenuItem */
+		$administer = $default->get('administer');
+		if (!$administer instanceof \ElggMenuItem || empty($administer->getChildren())) {
+			return null;
+		}
+		
+		/* @var $users \ElggMenuItem */
+		$users = null;
+		foreach ($administer->getChildren() as $child) {
+			if ($child->getID() === 'users') {
+				$users = $child;
+				break;
+			}
+		}
+		
+		if (!$users instanceof \ElggMenuItem || empty($users->getChildren())) {
+			return null;
+		}
+		
+		$children = $users->getChildren();
+		
+		$selected = $users->getSelected();
+		array_unshift($children, \ElggMenuItem::factory([
+			'name' => 'users:all',
+			'text' => elgg_echo('all'),
+			'href' => 'admin/users',
+			'parent_name' => 'users',
+			'priority' => 1,
+			'selected' => $selected,
+		]));
+		$users->setChildren($children);
+		
+		if ($selected) {
+			$users->addItemClass('elgg-has-selected-child');
+			$users->addItemClass('elgg-state-selected');
+		}
+		
+		$users->setHref(false);
+		
+		return $result;
+	}
+	
+	/**
+	 * Add the configure section to the admin page menu
+	 *
+	 * @param \Elgg\Event $event 'register', 'menu:admin_header'
+	 *
+	 * @return void|MenuItems
+	 */
+	public static function registerAdminConfigure(\Elgg\Event $event) {
+		if (!elgg_in_context('admin') || !elgg_is_admin_logged_in()) {
+			return;
+		}
+		
+		/* @var $return MenuItems */
+		$return = $event->getValue();
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'configure',
+			'text' => elgg_echo('menu:page:header:configure'),
+			'href' => false,
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'settings:basic',
+			'text' => elgg_echo('admin:site_settings'),
+			'href' => 'admin/site_settings',
+			'priority' => 10,
+			'parent_name' => 'configure',
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'settings:icons',
+			'text' => elgg_echo('admin:site_icons'),
+			'href' => 'admin/site_icons',
+			'priority' => 20,
+			'parent_name' => 'configure',
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'security',
+			'text' => elgg_echo('admin:security'),
+			'href' => 'admin/security',
+			'priority' => 30,
+			'parent_name' => 'configure',
+		]);
+		
+		// Utilities
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'configure_utilities',
+			'text' => elgg_echo('admin:configure_utilities'),
+			'href' => false,
+			'priority' => 600,
+			'parent_name' => 'configure',
+		]);
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'configure_utilities:maintenance',
+			'text' => elgg_echo('admin:configure_utilities:maintenance'),
+			'href' => 'admin/configure_utilities/maintenance',
+			'parent_name' => 'configure_utilities',
+		]);
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'configure_utilities:menu_items',
+			'text' => elgg_echo('admin:configure_utilities:menu_items'),
+			'href' => 'admin/configure_utilities/menu_items',
+			'parent_name' => 'configure_utilities',
+		]);
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'configure_utilities:robots',
+			'text' => elgg_echo('admin:configure_utilities:robots'),
+			'href' => 'admin/configure_utilities/robots',
+			'parent_name' => 'configure_utilities',
+		]);
+		
+		return $return;
+	}
+	
+	/**
+	 * Register menu items for default widgets
+	 *
+	 * @param \Elgg\Event $event 'register', 'menu:admin_header'
+	 *
+	 * @return void|MenuItems
+	 */
+	public static function registerAdminDefaultWidgets(\Elgg\Event $event) {
+		if (!elgg_in_context('admin') || !elgg_is_admin_logged_in()) {
+			return;
+		}
+		
+		if (empty(elgg_trigger_event_results('get_list', 'default_widgets', [], []))) {
+			return;
+		}
+		
+		/* @var $return MenuItems */
+		$return = $event->getValue();
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'default_widgets',
+			'text' => elgg_echo('admin:configure_utilities:default_widgets'),
+			'href' => 'admin/configure_utilities/default_widgets',
+			'parent_name' => 'configure_utilities',
+		]);
+		
+		return $return;
+	}
+	
+	/**
+	 * Add the information section to the admin page menu
+	 *
+	 * @param \Elgg\Event $event 'register', 'menu:admin_header'
+	 *
+	 * @return void|MenuItems
+	 */
+	public static function registerAdminInformation(\Elgg\Event $event) {
+		if (!elgg_in_context('admin') || !elgg_is_admin_logged_in()) {
+			return;
+		}
+		
+		/* @var $return MenuItems */
+		$return = $event->getValue();
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'information',
+			'text' => elgg_echo('menu:page:header:information'),
+			'href' => false,
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'server',
+			'text' => elgg_echo('admin:server'),
+			'href' => 'admin/server',
+			'parent_name' => 'information',
+			'priority' => 50,
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'information:security',
+			'text' => elgg_echo('admin:security'),
+			'href' => 'admin/security/information',
+			'parent_name' => 'information',
+			'priority' => 60,
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'information:performance',
+			'text' => elgg_echo('admin:performance'),
+			'href' => 'admin/performance',
+			'parent_name' => 'information',
+			'priority' => 70,
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'statistics',
+			'text' => elgg_echo('admin:statistics'),
+			'href' => 'admin/statistics',
+			'parent_name' => 'information',
+			'priority' => 80,
+		]);
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'cron',
+			'text' => elgg_echo('admin:cron'),
+			'href' => 'admin/cron',
+			'parent_name' => 'information',
+			'priority' => 90,
 		]);
 		
 		return $return;
