+++ mod/profile/classes/Elgg/Profile/Menus/AdminHeader.php
@@ -5,7 +5,7 @@ namespace Elgg\Profile\Menus;
 /**
  * Event callbacks for menus
  *
- * @since 4.0
+ * @since 5.0
  * @internal
  */
 class AdminHeader {
@@ -13,7 +13,7 @@ class AdminHeader {
 	/**
 	 * Register menu items for the topbar menu
 	 *
-	 * @param \Elgg\Event $event 'register' 'menu:topbar'
+	 * @param \Elgg\Event $event 'register' 'menu:admin_header'
 	 *
 	 * @return void|\Elgg\Menu\MenuItems
 	 */
@@ -31,7 +31,33 @@ class AdminHeader {
 			'text' => elgg_echo('profile'),
 			'icon' => 'user',
 			'parent_name' => 'account',
+			'section' => 'alt',
 			'priority' => 500,
+		]);
+		
+		return $return;
+	}
+	
+	/**
+	 * Register menu items for the admin page menu
+	 *
+	 * @param \Elgg\Event $event 'register' 'menu:admin_header'
+	 *
+	 * @return void|\Elgg\Menu\MenuItems
+	 */
+	public static function registerAdminProfileFields(\Elgg\Event $event) {
+		
+		if (!elgg_in_context('admin') || !elgg_is_admin_logged_in()) {
+			return;
+		}
+		
+		$return = $event->getValue();
+		
+		$return[] = \ElggMenuItem::factory([
+			'name' => 'configure_utilities:profile_fields',
+			'text' => elgg_echo('admin:configure_utilities:profile_fields'),
+			'href' => 'admin/configure_utilities/profile_fields',
+			'parent_name' => 'configure_utilities',
 		]);
 		
 		return $return;
