+++ mod/profile/elgg-plugin.php
@@ -60,9 +60,9 @@ return [
 		'register' => [
 			'menu:admin_header' => [
 				'Elgg\Profile\Menus\AdminHeader::register' => [],
+				'Elgg\Profile\Menus\AdminHeader::registerAdminProfileFields' => [],
 			],
 			'menu:page' => [
-				'Elgg\Profile\Menus\Page::registerAdminProfileFields' => [],
 				'Elgg\Profile\Menus\Page::registerProfileEdit' => [],
 			],
 			'menu:title' => [
