+++ mod/reportedcontent/classes/Elgg/ReportedContent/Menus/AdminHeader.php
@@ -5,24 +5,20 @@ namespace Elgg\ReportedContent\Menus;
 /**
  * Event callbacks for menus
  *
- * @since 4.0
+ * @since 5.0
  * @internal
  */
-class Page {
+class AdminHeader {
 	
 	/**
 	 * Add report user link to hover menu
 	 *
-	 * @param \Elgg\Event $event 'register', 'menu:page'
+	 * @param \Elgg\Event $event 'register', 'menu:admin_header'
 	 *
 	 * @return void|\Elgg\Menu\MenuItems
 	 */
 	public static function register(\Elgg\Event $event) {
-		if (!elgg_is_admin_logged_in()) {
-			return;
-		}
-		
-		if (!elgg_in_context('admin')) {
+		if (!elgg_in_context('admin') || !elgg_is_admin_logged_in()) {
 			return;
 		}
 				
@@ -31,7 +27,6 @@ class Page {
 			'name' => 'administer_utilities:reportedcontent',
 			'text' => elgg_echo('admin:administer_utilities:reportedcontent'),
 			'href' => 'admin/administer_utilities/reportedcontent',
-			'section' => 'administer',
 			'parent_name' => 'administer_utilities',
 		]);
